refactor(premium-analytics): share useElementSize between donut and semi-circle charts
