+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useAttributesWithSearchFallback } from './use-attributes-with-search-fallback';
 export { useChartTheme, type WooChartTheme } from './use-chart-theme';
+export { useElementSize, type ElementSize } from './use-element-size';
 export { useSegmentStyles } from '../widgets/common';
 export { useSeriesStyles } from './use-series-styles';
 export { useWidgetError } from './use-widget-error';
